Land rover with coordinate and orientation args
When rover lands the coordinates set the rover's position. This has been set up as an instance variable so that it can be accessed when functions are introduced to move the rover, and change the position. The same is true of orientation. They are set up as attr_readers so that they can be read from the outside, but can only be altared through Rover' methods.