Add orientator class
Orientator class has been added. This has been done in order to facilitate a more complex 'move' method in the Rover class, which currently just adds 1 to the position of the y axis, only fitting the brief when the rover is facing North. The orientator class can convert a compass point - North, East, South, West - to a bearing. Now, the rover class can interact with a bearing directly, which will allowing trigonometry functions to move rover can be used.