Changed rove 'move' method to take into account orientation
Move now functions uses triginometry to interact with bearings in radians. The return of cos(bearing) and sin(bearing) is either -1, 0, or 1. This is then multiplied with the ONE_MOVEMENT constant(currently 1) to be added to the x or y coordinate. The ONE_MOVEMENT constant has been implemented so that it is easy to change.